+++ test/Events.t.sol
@@ -4,16 +4,16 @@ pragma solidity ^0.8.13;
 import "forge-std/Test.sol";
 import "../src/Events.sol";
 
-contract Eventtest is Test{
+contract Eventtest is Test {
     Event public eve;
 
-    event Transfer(address indexed from, address indexed to, uint amount);
+    event Transfer(address indexed from, address indexed to, uint256 amount);
 
-    function setUp()public{
+    function setUp() public {
         eve = new Event();
     }
 
-    function testEmitTransferEvent()public{
+    function testEmitTransferEvent() public {
         vm.expectEmit(true, true, false, true);
         emit Transfer(address(this), address(123), 456);
         eve.transfer(address(this), address(123), 456);
@@ -24,17 +24,16 @@ contract Eventtest is Test{
         */
     }
 
-
-    function testEmitTransferMany()public{
+    function testEmitTransferMany() public {
         address[] memory to = new address[](2);
         to[0] = address(1783);
         to[1] = address(8907);
 
-        uint[]memory amounts = new uint256[](2);
+        uint256[] memory amounts = new uint256[](2);
         amounts[0] = 7890;
-        amounts[1] = 6784; 
+        amounts[1] = 6784;
 
-        for(uint i; i < to.length; i++){
+        for (uint256 i; i < to.length; i++) {
             vm.expectEmit(true, true, false, true);
             emit Transfer(address(this), to[i], amounts[i]);
         }
